@@ -4,7 +4,7 @@ version: 0.0.7
 homepage: https://github.com/jaumard/jimtl
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.4.4 <4.0.0"
 
 dependencies:
   intl: ">=0.17.0 <1.0.0"
