@@ -4,10 +4,10 @@ version: 0.0.4
 homepage: https://github.com/jaumard/jimtl
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.4.4 <4.0.0"
 
 dependencies:
-  analyzer: ">=5.0.0 <6.0.0"
+  analyzer: ">=5.0.0 <8.0.0"
   build: ">=2.0.1 <3.0.0"
   path: ">=1.8.0 <2.0.0"
   source_gen: ">=1.0.0 <2.0.0"
